graph admin

+++ app/Http/Controllers/Admin/AdminDashboardController.php
@@ -46,7 +46,7 @@ class AdminDashboardController extends Controller
 
             $monthlySales[] = [
                 'month' => $month->format('M'),
-                'amount' => $sales ?? 0
+                'amount' => (int) ($sales ?? 0)
             ];
         }
